Add support for vhostuser with empty dir support with libvirt
userspace cni can be used with kubevirt, which can create the
vhostuser socket in the server mode and ovs can act as a client.
An empty directory can be used in the libvirt container, which
create vhostuser socket in a specific kubelet directory. Since
the directory length exceeds the limit 108 characheters, this
directory is mounted to a local directory in order to shorten
the vhostuser path added to ovs.

Signed-off-by: Saravanan KR <skramaja@redhat.com>